+++ EssentialFeed/EssentialFeedTests/Feed Api/URLSessionHTTPClientTests.swift
@@ -142,10 +142,6 @@ final class URLSessionHTTPClientTests: XCTestCase {
         Data("any data".utf8)
     }
     
-    private func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 0)
-    }
-    
     private func nonHTTPURLResponse() -> URLResponse {
         URLResponse(url: anyURL(), mimeType: nil, expectedContentLength: 0, textEncodingName: nil)
     }
